refactored code and tests
