+++ requirements-tests.txt
@@ -1,6 +1,6 @@
 -r requirements.txt
-nose==1.3.4
-mock==1.1.3
-coverage==3.7.1
+pytest==2.7.2
+pytest-cov==2.0.0
+pytest-pep8==1.0.6
 python-coveralls==2.5.0
-
+mock==1.1.3
